Use DCE library on HP-UX to handle UUID generation
This is the initial dumb implementation - it just replaces RAND_bytes
in the final uuid case rather than handling uuid_create() as done in
the BSD case.

It passes the self tests, but it's probably not the correct way to do
it.

You can also use the openssl method my using the '--without-dce' flag
to configure.

Signed-off-by: AJ Lewis <aj.lewis@quantum.com>