Change peer_identity_size (in all cases) in the command_t structure to be a size_t instead of an unsigned char.
The peer_identity_size field in the attach and bind substructure is
used to generate a C++ base_string object from an unsigned char *, and
that object expects a size_t, not an unsigned char.  If an unsigned
char is used, the HPUX aCC compiler gets confused and can't figure out
which constructor to use.

Signed-off-by: AJ Lewis <aj.lewis@quantum.com>