Fix bug in connecter
The connecter used EAGAIN to signal that the connection
couldn't be completed immediately. But the same value was
used by connect(2) to signal that it failed. This made it
impossible for the connecter to distinguish between these
conditions.

The patch modifies the connecter to use EINPROGRESS value,
as originally returned by connect(2).

Another problem addressed by the patch is the situation
when a connection could not be completed immediately
with the IPC transport. The patch modifies the connecter
so that it behaves similarly to TCP transport.

Refs. https://zeromq.jira.com/browse/LIBZMQ-381